pyproject.toml
[tool.poetry]
name = "telegram-star-bot"
version = "0.1.0"
description = ""
authors = ["sepehr masihpour"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.2.1"
pydantic = "^2.12.3"
httpx = "^0.28.1"
fastapi = "^0.120.4"
pydantic-settings = "^2.11.0"
pyngrok = "^7.4.1"
uvicorn = "^0.38.0"
sqlalchemy = "^2.0.44"
psycopg = {extras = ["binary"], version = "^3.2.12"}
python-multipart = "^0.0.21"
argon2-cffi = "^25.1.0"
pyotp = "^2.9.0"
pyjwt = "^2.11.0"


[tool.poetry.group.dev.dependencies]
alembic = "^1.17.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
